Split config.h into AXI4 and Target configs
The target config relies on the optional type, which cannot be compiled using
verilator's fixed C++ standard version. This PR splits the problematic header
into two, so the files which must be placed under Verilator can include headers
that do not rely on any recent C++ features.
